Issue #16: Integrated UserCache to UserCacheAdapter.
